// site/types/contentful.generated.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IBlogFields {
  /** タイトル */
  title?: string | undefined;

  /** 本文 */
  body?: string | undefined;

  /** カバー画像 */
  coverImage?: Asset | undefined;
}

export interface IBlog extends Entry<IBlogFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "blog";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IExternalFields {
  /** タイトル */
  title?: string | undefined;

  /** 埋め込みコンテンツURL */
  url: string;
}

export interface IExternal extends Entry<IExternalFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "external";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageFields {
  /** ID */
  id: string;

  /** タイトル */
  title: string;

  /** 本文 */
  body: string;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IWorkFields {
  /** タイトル */
  title: string;

  /** スラッグ */
  slug: string;

  /** キャッチコピー */
  copyText: string;

  /** リード文 */
  leadText: string;

  /** 本文 */
  body: string;

  /** カバー画像 */
  coverImage: Asset;

  /** 制作開始日 */
  dateStart: string;

  /** 制作終了日 */
  dateEnd?: string | undefined;

  /** ギャラリー */
  gallery?: Asset[] | undefined;

  /** 順序 */
  order?: number | undefined;
}

export interface IWork extends Entry<IWorkFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "work";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "blog" | "external" | "page" | "work";

export type LOCALE_CODE = "en-US" | "ja";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "ja";